requirements.txt
# requirements.txt

torch>=1.12.0
torchvision>=0.13.0
ftfy
regex
clip @ git+https://github.com/openai/CLIP.git
PyYAML>=5.4
numpy>=1.19
pillow>=8.0
scikit-image>=0.18
scipy>=1.6
tqdm>=4.60
pandas>=1.2
matplotlib>=3.3
